+++ Tool1/TraceGenerator/Packetloss/PacketlossParameterParser/BruteForceParser.h
@@ -10,16 +10,20 @@ using namespace std;
 
 class BruteForceParser {
 protected:
-    void calcLoss(vector<bool> &trace, float &lossrate, float &burstsize, float &goodsize, vector<int> &overallsizes);
+    void
+    calcLoss(vector<bool> &trace, double &lossrate, double &burstsize, double &goodsize, vector<int> &overallsizes);
 
-    void calcDistFunction(vector<int> &sizes, vector<vector<float> > &distFunction);
+    void calcDistFunction(vector<int> &sizes, vector<vector<double> > &distFunction);
 
-    bool kstest(vector<vector<float> > origDistFunction, vector<vector<float> > &generatedDistFunction, long m, long n, float &dvalue);
+    bool
+    kstest(vector<vector<double> > origDistFunction, vector<vector<double> > &generatedDistFunction, long m, long n,
+           double &dvalue);
 
-    void findTopX(vector<vector<float> > &top, vector<vector<float> > &possibleParams, int x, string model);
+    void findTopX(vector<vector<double> > &top, vector<vector<double> > &possibleParams, int x, string model);
 
-    virtual vector<float> checkResult(vector<vector<float>> avgburstsizes, float origburstsize, vector<vector<float>> possibleParams,
-                float origLoss, float origgoodsize);
+    virtual vector<double>
+    checkResult(vector<vector<double>> avgburstsizes, double origburstsize, vector<vector<double>> possibleParams,
+                double origLoss, double origgoodsize);
 };
